+++ tests/test_maildir_backend.rs
@@ -21,6 +21,10 @@ fn test_maildir_backend() {
     if let Err(_) = fs::remove_dir_all(mdir_sub.path()) {}
     mdir_sub.create_dirs().unwrap();
 
+    let mdir_trash = Maildir::from(mdir.path().join(".Trash"));
+    if let Err(_) = fs::remove_dir_all(mdir_trash.path()) {}
+    mdir_trash.create_dirs().unwrap();
+
     let account_config = AccountConfig {
         name: "account".into(),
         folder_aliases: HashMap::from_iter([("subdir".into(), "Subdir".into())]),
@@ -113,29 +117,53 @@ fn test_maildir_backend() {
         .unwrap();
     let inbox = mdir.list_envelopes("INBOX", 0, 0).unwrap();
     let subdir = mdir.list_envelopes("subdir", 0, 0).unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
     assert_eq!(1, inbox.len());
     assert_eq!(1, subdir.len());
-    assert!(mdir.get_emails("INBOX", vec![&id]).is_ok());
-    assert!(mdir.get_emails("subdir", vec![&id]).is_ok());
-    assert!(submdir.get_emails("INBOX", vec![&id]).is_ok());
+    assert_eq!(0, trash.len());
+    assert!(mdir.get_emails("INBOX", vec![&inbox[0].id]).is_ok());
+    assert!(mdir.get_emails("subdir", vec![&subdir[0].id]).is_ok());
+    assert!(submdir.get_emails("INBOX", vec![&subdir[0].id]).is_ok());
+
+    // check that the email can be marked as deleted then expunged
+    mdir.mark_emails_as_deleted("subdir", vec![&subdir[0].id])
+        .unwrap();
+    let inbox = mdir.list_envelopes("INBOX", 0, 0).unwrap();
+    let subdir = mdir.list_envelopes("subdir", 0, 0).unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
+    assert_eq!(1, inbox.len());
+    assert_eq!(1, subdir.len());
+    assert_eq!(0, trash.len());
+
+    mdir.expunge_folder("subdir").unwrap();
+    let subdir = mdir.list_envelopes("subdir", 0, 0).unwrap();
+    assert_eq!(0, subdir.len());
 
     // check that the message can be moved
     mdir.move_emails("INBOX", "subdir", vec![&envelope.id])
         .unwrap();
     let inbox = mdir.list_envelopes("INBOX", 0, 0).unwrap();
     let subdir = mdir.list_envelopes("subdir", 0, 0).unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
     assert_eq!(0, inbox.len());
     assert_eq!(1, subdir.len());
-    assert!(mdir.get_emails("INBOX", vec![&id]).is_err());
-    assert!(mdir.get_emails("subdir", vec![&id]).is_ok());
-    assert!(submdir.get_emails("INBOX", vec![&id]).is_ok());
+    assert_eq!(0, trash.len());
 
     // check that the message can be deleted
-    mdir.delete_emails("subdir", vec![&id]).unwrap();
+    mdir.delete_emails("subdir", vec![&subdir[0].id]).unwrap();
     let inbox = mdir.list_envelopes("INBOX", 0, 0).unwrap();
     let subdir = mdir.list_envelopes("subdir", 0, 0).unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
     assert_eq!(0, inbox.len());
     assert_eq!(0, subdir.len());
-    assert!(mdir.get_emails("subdir", vec![&id]).is_err());
-    assert!(submdir.get_emails("INBOX", vec![&id]).is_err());
+    assert_eq!(1, trash.len());
+
+    mdir.delete_emails("Trash", vec![&trash[0].id]).unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
+    assert_eq!(1, trash.len());
+    assert!(trash[0].flags.contains(&Flag::Deleted));
+
+    mdir.expunge_folder("Trash").unwrap();
+    let trash = mdir.list_envelopes("Trash", 0, 0).unwrap();
+    assert_eq!(0, trash.len());
 }