+++ tests/test_sync.rs
@@ -42,6 +42,7 @@ fn test_sync() {
     // set up folders
 
     imap.add_folder("[Gmail]/Sent").unwrap();
+    imap.add_folder("Trash").unwrap();
 
     // add three emails to folder INBOX with delay (in order to have a
     // different date)
@@ -261,6 +262,7 @@ fn test_sync() {
         &Flags::from_iter([Flag::Draft]),
     )
     .unwrap();
+    imap.expunge_folder("INBOX").unwrap();
     mdir.delete_emails_internal("INBOX", vec![&mdir_inbox_envelopes[2].internal_id])
         .unwrap();
     mdir.add_flags_internal(
@@ -269,11 +271,12 @@ fn test_sync() {
         &Flags::from_iter([Flag::Flagged, Flag::Answered]),
     )
     .unwrap();
+    mdir.expunge_folder("INBOX").unwrap();
 
     let report = sync_builder.sync(&imap).unwrap();
     assert_eq!(
         report.folders,
-        HashSet::from_iter(["INBOX".into(), "[Gmail]/Sent".into()])
+        HashSet::from_iter(["INBOX".into(), "[Gmail]/Sent".into(), "Trash".into()])
     );
 
     let imap_envelopes = imap.list_envelopes("INBOX", 0, 0).unwrap();
@@ -292,5 +295,6 @@ fn test_sync() {
 
     imap.purge_folder("INBOX").unwrap();
     imap.delete_folder("[Gmail]/Sent").unwrap();
+    imap.delete_folder("Trash").unwrap();
     imap.close().unwrap();
 }